+++ lerna.json
@@ -4,6 +4,7 @@
     "packages/tests",
     "packages/benchmarks-utils",
     "packages/dev-tasks",
+    "packages/benchmarks/use-style",
     "packages/benchmarks/fela",
     "packages/benchmarks/react-native-web",
     "packages/benchmarks/rockey/rockey",
@@ -35,8 +36,7 @@
     "packages/benchmarks/styled-jsx/inline-styles",
     "packages/benchmarks/styled-jsx/dynamic",
     "packages/benchmarks/styletron",
-    "packages/benchmarks/reshadow",
-    "packages/benchmarks/css-system"
+    "packages/benchmarks/reshadow"
   ],
   "version": "1.0.0"
 }
