+++ packages/benchmarks/use-style/client/index.html
@@ -1,8 +1,8 @@
-<!doctype html>
+<!DOCTYPE html>
 <html>
   <head>
-    <meta charset="utf-8">
-    <title>React</title>
+    <meta charset="utf-8" />
+    <title>use-style</title>
   </head>
   <body>
     <div id="root"></div>
